jwToken

+++ router/router.php
@@ -19,7 +19,7 @@ if ($rMethod === "GET") {
         $cont->badRequestJsonResponse();
     }
 } elseif ($rMethod === "POST") {
-    if (isset($uri[2]) && $uri[2] === "create") {
+    if (isset($uri[2]) && $uri[2] === "create" && in_array("author", $role)) {
         $cont->create($_POST);
     } elseif (isset($uri[2]) && $uri[1] === "user" && $uri[2] === "login") {
         $cont->login($_POST);
@@ -30,7 +30,7 @@ if ($rMethod === "GET") {
     }
 } elseif ($rMethod === "PUT" || $rMethod === "PATCH") {
     if (isset($uri[2]) && preg_match("[\d]", $uri[2])) {
-        if (isset($uri[3]) && $uri[3] === "update") {
+        if (isset($uri[3]) && $uri[3] === "update" && in_array("author", $role)) {
             $_PUT = array();
             parse_str(file_get_contents("php://input"), $_PUT);
             $cont->update($uri[2], $_PUT);
@@ -42,7 +42,7 @@ if ($rMethod === "GET") {
     }
 } elseif ($rMethod === "DELETE") {
     if (isset($uri[2]) && preg_match("[\d]", $uri[2])) {
-        if (isset($uri[3]) && $uri[3] === "delete") {
+        if (isset($uri[3]) && $uri[3] === "delete" && in_array("author", $role)) {
             $cont->delete($uri[2]);
         } else {
             $cont->notAllowedResponse("Page not found please try ".SERVER .$uri[1]."/{id}/delete with method DELETE.");
